change Manager cunstructor
