+++ Phase1/classes/Main.java
@@ -5,9 +5,13 @@
  */
 package classes;
 
+import java.util.Scanner;
 import java.util.Set;
 
 public class Main {
+
+    private static final Scanner scanner = new Scanner(System.in);
+
     /**
      * the main method of the application this method initializes the document data
      * that we stored in Docs folder and run the application in an infinite loop for
@@ -19,7 +23,15 @@ public class Main {
         Manager.initialize(); // initializing the documents and making them ready for invertedIndexSearch
         while (true) { // inifite loop for taking infinite tasks from user
             System.out.println("Enter -1 when you want to finish the process.");
-            Manager manager = new Manager();
+
+            Manager manager = new Manager(new MyScanner(){
+                public String takeInput() {
+                    System.out.println("Enter the phrase to search for");
+                    final String input = scanner.nextLine();
+                    return input;
+                }
+            });
+            
             Set<String> output = manager.run(); // managing the input from user.
             for (String doc : output) {
                 System.out.println(doc);
